+++ backend/ModernWMS.WMS/Entities/ViewModels/Stockadjust/StockadjustViewModel.cs
@@ -128,7 +128,7 @@ namespace ModernWMS.WMS.Entities.ViewModels
         /// tenant_id
         /// </summary>
         [Display(Name = "tenant_id")]
-        public byte tenant_id { get; set; }  = 0;
+        public long tenant_id { get; set; }  = 0;
 
         /// <summary>
         /// is_update_stock
